new effects

+++ src/main/java/com/solegendary/reignofnether/unit/units/monsters/HuskUnit.java
@@ -158,7 +158,7 @@ public class HuskUnit extends Husk implements Unit, AttackerUnit {
     }
 
     @Override
-    public float getUnitRangedArmorPercentage() {
+    public double getUnitRangedArmorPercentage() {
         return rangedDamageResist;
     }
