+++ src/main/java/com/solegendary/reignofnether/unit/units/villagers/VindicatorUnit.java
@@ -164,7 +164,7 @@ public class VindicatorUnit extends Vindicator implements Unit, AttackerUnit {
     }
 
     @Override
-    public float getUnitRangedArmorPercentage() {
+    public double getUnitRangedArmorPercentage() {
         return rangedDamageResist;
     }
